test: split oversized skill contracts
