+++ scripts/test/safe-core-files.json
@@ -97,6 +97,7 @@
   "tests/skills/home-status-redesign-contract.test.ts",
   "tests/skills/honesty-and-gate-pins.test.ts",
   "tests/skills/indirect-actuation-contract.test.ts",
+  "tests/skills/indirect-actuation-expansion-contract.test.ts",
   "tests/skills/integration-setup-contract.test.ts",
   "tests/skills/maintenance-contract.test.ts",
   "tests/skills/output-design-contract.test.ts",
@@ -108,6 +109,7 @@
   "tests/skills/service-call-contract.test.ts",
   "tests/skills/skill-source-language-contract.test.ts",
   "tests/skills/skill-template-contract.test.ts",
+  "tests/skills/skill-word-budget-contract.test.ts",
   "tests/skills/smallest-solution-contract.test.ts",
   "tests/skills/test-run-contract.test.ts",
   "tests/skills/threshold-calibration-contract.test.ts",
